Split node into modules
